Avoid running any SQL using plpy.execute
